requirements.txt
pyTMCL~=1.0.10
pyserial~=3.5